+++ qjazz-server/qjazz-map/src/handlers/conformance.rs
@@ -6,8 +6,8 @@ use actix_web::{HttpRequest, HttpResponse};
 
 use serde::Serialize;
 
-use crate::handlers::utils::request;
 use crate::models::{Link, rel};
+use crate::requests::request;
 
 #[derive(Debug, Serialize)]
 #[serde(rename_all = "camelCase")]