+++ qjazz-server/qjazz-map/src/handlers/landing_page.rs
@@ -2,8 +2,8 @@ use actix_web::{HttpRequest, HttpResponse, Responder, web};
 use serde::Serialize;
 
 use crate::channel::Channel;
-use crate::handlers::utils::request;
 use crate::models::{Link, rel};
+use crate::requests::request;
 //use crate::resolver::ApiEndPoint;
 
 type Channels = Vec<web::Data<Channel>>;