+++ qjazz-server/qjazz-map/src/main.rs
@@ -1,3 +1,4 @@
+mod admin;
 mod channel;
 mod config;
 mod cors;
@@ -5,7 +6,9 @@ mod handlers;
 mod logger;
 mod models;
 mod monitor;
+mod requests;
 mod resolver;
+mod responses;
 mod server;
 mod services;
 mod utils;